# apps/next/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datasets Generator API",
    "description": "API to generate datasets from web scraping",
    "version": "0.0.1"
  },
  "paths": {
...
    "/q_a/{dataset_name}/stats": {
      "get": {
        "tags": [
          "q_a"
        ],
        "summary": "Get Qa Stats",
        "description": "Aggregated confidence-score stats over the whole dataset (server-side).\n\nAvoids the client having to page through /q_a/{dataset_name} to compute\naverages and distributions — accurate even on large datasets.",
        "operationId": "get_qa_stats_q_a__dataset_name__stats_get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          },
          {
            "name": "score_threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Validated/below split threshold",
              "default": 0.8,
              "title": "Score Threshold"
            },
            "description": "Validated/below split threshold"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/q_a/{dataset_name}/score": {
      "post": {
        "tags": [
          "q_a"
        ],
        "summary": "Score Qa",
        "description": "Score a dataset's pairs with an LLM judge and store the confidences.\n\nFor pairs that have no score yet — typically ones imported from the Hub.\nWith a context, the judge rates how well it supports the answer (the same\nmeasure generation reports); without one, the pair's own quality.",
        "operationId": "score_qa_q_a__dataset_name__score_post",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          },
          {
            "name": "only_unscored",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Score only pairs without a confidence (false: rescore all)",
              "default": true,
              "title": "Only Unscored"
            },
            "description": "Score only pairs without a confidence (false: rescore all)"
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Judge model reference (defaults to your qa model)",
              "title": "Model"
            },
            "description": "Judge model reference (defaults to your qa model)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/q_a/{dataset_name}": {
      "get": {
        "tags": [
          "q_a"
        ],
        "summary": "Get Qa By Dataset",
        "description": "Retrieve a dataset's Q&A pairs (keyed by dataset name).",
        "operationId": "get_qa_by_dataset_q_a__dataset_name__get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limit number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Limit number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QAItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique ID of the question-answer"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "description": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Answer"
          },
          "context": {
            "type": "string",
            "title": "Context",
            "description": "Source context"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "Source URL"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Confidence level (null when unscored)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation date"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Additional metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question",
          "answer",
          "context",
          "created_at"
        ],
        "title": "QAItem",
        "description": "Model for an individual Q&A item"
      },
      "QAListResponse": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Dataset name"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset ID"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of items"
          },
          "returned_count": {
            "type": "integer",
            "title": "Returned Count",
            "description": "Number of items returned"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Applied offset",
            "default": 0
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit",
            "description": "Applied limit"
          },
          "qa_data": {
            "items": {
              "$ref": "#/components/schemas/QAItem"
            },
            "type": "array",
            "title": "Qa Data",
            "description": "List of question-answers"
          }
        },
        "type": "object",
        "required": [
          "dataset_name",
          "dataset_id",
          "total_count",
          "returned_count",
          "qa_data"
        ],
        "title": "QAListResponse",
        "description": "Response model for a dataset's Q&A list"
      },
...
      "QAScoreBucket": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Human-readable bucket range, e.g. '0.9–1.0'"
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count",
            "description": "Number of scored items in the bucket"
          }
        },
        "type": "object",
        "required": [
          "label",
          "count"
        ],
        "title": "QAScoreBucket",
        "description": "One bucket of the confidence-score distribution"
      },
      "QAScoreResponse": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Dataset name"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model used as the judge"
          },
          "requested": {
            "type": "integer",
            "title": "Requested",
            "description": "Pairs sent to the judge"
          },
          "scored": {
            "type": "integer",
            "title": "Scored",
            "description": "Pairs that received a score"
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "description": "Pairs left unscored (judge error/omission)"
          }
        },
        "type": "object",
        "required": [
          "dataset_name",
          "model",
          "requested",
          "scored",
          "failed"
        ],
        "title": "QAScoreResponse",
        "description": "Result of scoring a dataset's pairs with the LLM judge"
      },
      "QAStatsResponse": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Dataset name"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset ID"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of active items"
          },
          "scored_count": {
            "type": "integer",
            "title": "Scored Count",
            "description": "Items carrying a confidence score (the stats basis)"
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score",
            "description": "Mean confidence of scored items (null when none)"
          },
          "score_threshold": {
            "type": "number",
            "title": "Score Threshold",
            "description": "Threshold used for the split"
          },
          "below_threshold_count": {
            "type": "integer",
            "title": "Below Threshold Count",
            "description": "Scored items strictly below the threshold"
          },
          "validated_count": {
            "type": "integer",
            "title": "Validated Count",
            "description": "Scored items at or above the threshold"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/QAScoreBucket"
            },
            "type": "array",
            "title": "Distribution",
            "description": "Score distribution buckets, highest range first"
          }
        },
        "type": "object",
        "required": [
          "dataset_name",
          "dataset_id",
          "total_count",
          "scored_count",
          "score_threshold",
          "below_threshold_count",
          "validated_count",
          "distribution"
        ],
        "title": "QAStatsResponse",
        "description": "Aggregated confidence-score statistics over a whole dataset.\n\nComputed server-side over every active item so the quality page doesn't\nhave to sample a capped page of Q&A items client-side."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}